IsNullOrWhiteSpace method is not supported on PS 2.0
We do not support IsNullOrWhiteSpace method on PowerShell 2.0.  Without this we're not able to use the New-LoggerObject function which is part of the new default debug log. We have two option here:
1) Re-write the New-LoggerObject function to work properly on PowerShell 2.0
2) Silently continue if method is not supported

I've added some try / catch blocks to silently continue without writing debug log as a quick fix for this issue. We should decide if we want to make the New-LoggerObject function working on PS 2.0 or not. Extended support for  Server 2008 R2 SP1 will end on 1/14/2020.